Fixed issue with seed
When doing a model, a seed would not be specified and the command would default to 12345. When running tests on beta distribution, the betas from the model would be constant. This only affects model stage, if running models within a loop where any sort of random number generation plays a role. It has no bearing on simulation results or others.